# chat-backend/docs/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "TAI后端服务",
    "title": "TAI Backend API",
    "version": "1.0.0"
  },
  "host": "localhost:9090",
  "basePath": "/api/v1",
  "paths": {
...
    "/knowledge/bases": {
      "get": {
        "description": "获取系统中所有已创建的知识库列表，包括知识库基本信息",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Knowledge"
        ],
        "summary": "获取知识库列表",
        "operationId": "listKnowledgeBases",
        "responses": {
          "200": {
            "$ref": "#/responses/models.KnowledgeBase"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      },
      "post": {
        "description": "创建一个新的知识库，用于存储和管理相关文档",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Knowledge"
        ],
        "summary": "创建知识库",
        "operationId": "createKnowledgeBase",
        "parameters": [
          {
            "description": "知识库信息",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "description": "知识库信息",
              "type": "object",
              "$ref": "#/definitions/KnowledgeBaseCreateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/MessageWithDataResponse"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/knowledge/bases/{id}": {
      "put": {
        "description": "更新指定知识库的名称、描述等基本信息",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Knowledge"
        ],
        "summary": "更新知识库",
        "operationId": "updateKnowledgeBase",
        "parameters": [
          {
            "type": "integer",
            "description": "知识库ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "更新信息",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "description": "更新信息",
              "type": "object",
              "$ref": "#/definitions/UpdateKnowledgeBaseRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/MessageWithDataResponse"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      },
      "delete": {
        "description": "根据知识库ID删除指定的知识库及其所有相关文件",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Knowledge"
        ],
        "summary": "删除知识库",
        "operationId": "deleteKnowledgeBase",
        "parameters": [
          {
            "type": "integer",
            "description": "知识库ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/MessageOnlyResponse"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/knowledge/bases/{id}/files": {
      "get": {
        "description": "获取指定知识库中所有已上传文件的列表和状态信息",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Knowledge"
        ],
        "summary": "获取知识库文件",
        "operationId": "getKnowledgeBaseFiles",
        "parameters": [
          {
            "type": "integer",
            "description": "知识库ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/models.KnowledgeFile"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      },
      "post": {
        "description": "支持两种上传方式：\n文件流上传：使用 multipart/form-data 格式，字段名为 'file',多个文件使用多个 'file' 字段\n文件路径上传：使用 application/json 格式，body 中包含 file_paths 数组",
        "consumes": [
          "multipart/form-data",
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Knowledge"
        ],
        "summary": "上传文件到知识库",
        "operationId": "uploadFile",
        "parameters": [
          {
            "type": "integer",
            "description": "知识库ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "object",
            "description": "上传文件（支持单个或多个文件）",
            "name": "file"
          },
          {
            "description": "单文件路径或多文件路径列表（用于路径上传）",
            "name": "body",
            "in": "body",
            "schema": {
              "description": "单文件路径或多文件路径列表（用于路径上传）",
              "type": "object",
              "$ref": "#/definitions/BatchUploadFilesRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "BatchUploadResponse",
            "schema": {
              "$ref": "#/definitions/BatchUploadResponse"
            }
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/knowledge/files/{file_id}": {
      "delete": {
        "description": "从知识库中删除指定的文件及其向量数据",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Knowledge"
        ],
        "summary": "删除知识库文件",
        "operationId": "deleteFile",
        "parameters": [
          {
            "type": "integer",
            "description": "文件ID",
            "name": "file_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/MessageOnlyResponse"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/knowledge/files/{file_id}/toggle": {
      "put": {
        "description": "启用或禁用知识库中的指定文件，控制其是否参与问答检索",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Knowledge"
        ],
        "summary": "切换文件启用状态",
        "operationId": "toggleFileEnable",
        "parameters": [
          {
            "type": "integer",
            "description": "文件ID",
            "name": "file_id",
            "in": "path",
            "required": true
          },
          {
            "description": "状态切换",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "description": "状态切换",
              "type": "object",
              "$ref": "#/definitions/FileToggleEnableRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/MessageOnlyResponse"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "BatchUploadFilesRequest": {
      "description": "BatchUploadFilesRequest 批量上传文件请求",
      "type": "object",
      "required": [
        "file_paths"
      ],
      "properties": {
        "file_paths": {
          "description": "上传文件路径列表（本地文件系统路径）",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "FilePaths"
        }
      },
      "x-go-package": "chat-backend/models"
    },
    "BatchUploadResponse": {
      "description": "BatchUploadResponse 批量上传响应",
      "type": "object",
      "required": [
        "total",
        "success_count",
        "failure_count",
        "results"
      ],
      "properties": {
        "failure_count": {
          "description": "失败数",
          "type": "integer",
          "format": "int64",
          "x-go-name": "FailureCount"
        },
        "results": {
          "description": "每个文件的上传结果",
          "type": "array",
          "items": {
            "$ref": "#/definitions/BatchUploadResult"
          },
          "x-go-name": "Results"
        },
        "success_count": {
          "description": "成功数",
          "type": "integer",
          "format": "int64",
          "x-go-name": "SuccessCount"
        },
        "total": {
          "description": "上传总数",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Total"
        }
      },
      "x-go-package": "chat-backend/models"
    },
    "BatchUploadResult": {
      "description": "BatchUploadResult 单个文件上传结果",
      "type": "object",
      "required": [
        "file_path",
        "success",
        "message"
      ],
      "properties": {
        "error": {
          "description": "错误信息（失败时返回）",
          "type": "string",
          "x-go-name": "Error"
        },
        "file": {
          "$ref": "#/definitions/KnowledgeFile"
        },
        "file_path": {
          "description": "文件路径",
          "type": "string",
          "x-go-name": "FilePath"
        },
        "message": {
          "description": "上传结果信息",
          "type": "string",
          "x-go-name": "Message"
        },
        "success": {
          "description": "上传是否成功",
          "type": "boolean",
          "x-go-name": "Success"
        }
      },
      "x-go-package": "chat-backend/models"
    },
...
    "FileToggleEnableRequest": {
      "description": "FileToggleEnableRequest 文件启用/禁用请求",
      "type": "object",
      "required": [
        "enable"
      ],
      "properties": {
        "enable": {
          "description": "true=启用, false=禁用",
          "type": "boolean",
          "x-go-name": "Enable"
        }
      },
      "x-go-package": "chat-backend/models"
    },
...
    "KnowledgeBaseCreateRequest": {
      "description": "KnowledgeBaseCreateRequest 创建知识库请求",
      "type": "object",
      "required": [
        "name",
        "desc",
        "vectorModel",
        "agentModel",
        "chunkStrategy",
        "chunkSize"
      ],
      "properties": {
        "agentModel": {
          "description": "对话模型ID",
          "type": "integer",
          "format": "int64",
          "x-go-name": "AgentModel"
        },
        "chunkSize": {
          "description": "切片大小",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ChunkSize"
        },
        "chunkStrategy": {
          "description": "切片策略 固定尺寸:fixed  自然句:period   自然段落:paragraph",
          "type": "string",
          "x-go-name": "ChunkStrategy"
        },
        "desc": {
          "description": "知识库描述",
          "type": "string",
          "x-go-name": "Desc"
        },
        "name": {
          "description": "知识库名称",
          "type": "string",
          "x-go-name": "Name"
        },
        "vectorModel": {
          "description": "向量模型ID",
          "type": "integer",
          "format": "int64",
          "x-go-name": "VectorModel"
        }
      },
      "x-go-package": "chat-backend/models"
    },
...
    "KnowledgeFile": {
      "description": "KnowledgeFile 知识库文件",
      "type": "object",
      "required": [
        "id",
        "name",
        "size",
        "enable",
        "status",
        "uploaded_at",
        "index_percent",
        "errorMessage"
      ],
      "properties": {
        "enable": {
          "description": "启用状态: true=启用, false=禁用",
          "type": "boolean",
          "x-go-name": "Enable"
        },
        "errorMessage": {
          "description": "错误信息，空字符串表示无错误",
          "type": "string",
          "x-go-name": "ErrorMessage"
        },
        "id": {
          "description": "文件ID",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "index_percent": {
          "description": "索引进度百分比",
          "type": "integer",
          "format": "int64",
          "x-go-name": "IndexPercent"
        },
        "name": {
          "description": "文件名称",
          "type": "string",
          "x-go-name": "Name"
        },
        "size": {
          "description": "文件大小（字节）",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Size"
        },
        "status": {
          "description": "索引状态: 0=构建中, 1=完成, 2=失败",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        },
        "uploaded_at": {
          "description": "上传时间",
          "type": "string",
          "format": "date-time",
          "x-go-name": "UploadedAt"
        }
      },
      "x-go-package": "chat-backend/models"
    },
...
    "ResponseBody": {
      "description": "ResponseBody 统一响应体结构",
      "type": "object",
      "required": [
        "is_login",
        "code",
        "msg",
        "data",
        "time",
        "code_num",
        "code_msg"
      ],
      "properties": {
        "code": {
          "description": "请求是否成功",
          "type": "boolean",
          "x-go-name": "Code"
        },
        "code_msg": {
          "description": "错误代码消息",
          "type": "string",
          "x-go-name": "CodeMsg"
        },
        "code_num": {
          "description": "数字错误代码",
          "type": "integer",
          "format": "int64",
          "x-go-name": "CodeNum"
        },
        "data": {
          "description": "响应数据",
          "x-go-name": "Data"
        },
        "is_login": {
          "description": "登录状态",
          "type": "boolean",
          "x-go-name": "IsLogin"
        },
        "msg": {
          "description": "响应消息",
          "type": "string",
          "x-go-name": "Message"
        },
        "time": {
          "description": "持续时间（秒）",
          "type": "number",
          "format": "double",
          "x-go-name": "Time"
        }
      },
      "x-go-package": "chat-backend/utils"
    },
...
    "UpdateKnowledgeBaseRequest": {
      "description": "UpdateKnowledgeBaseRequest 更新知识库请求",
      "type": "object",
      "required": [
        "name",
        "desc",
        "vectorModel",
        "agentModel",
        "chunkStrategy",
        "chunkSize"
      ],
      "properties": {
        "agentModel": {
          "description": "对话模型ID",
          "type": "integer",
          "format": "int64",
          "x-go-name": "AgentModel"
        },
        "chunkSize": {
          "description": "切片大小",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ChunkSize"
        },
        "chunkStrategy": {
          "description": "切片策略 固定尺寸:fixed  自然句:period   自然段落:paragraph",
          "type": "string",
          "x-go-name": "ChunkStrategy"
        },
        "desc": {
          "description": "知识库描述",
          "type": "string",
          "x-go-name": "Desc"
        },
        "name": {
          "description": "知识库名称",
          "type": "string",
          "x-go-name": "Name"
        },
        "vectorModel": {
          "description": "向量模型ID",
          "type": "integer",
          "format": "int64",
          "x-go-name": "VectorModel"
        }
      },
      "x-go-package": "chat-backend/models"
    },
...
  },
  "responses": {
...
    "MessageOnlyResponse": {
      "description": "MessageOnlyResponse 仅消息成功响应",
      "schema": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "description": "响应消息",
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    },
    "MessageWithDataResponse": {
      "description": "MessageWithDataResponse 消息和数据成功响应",
      "schema": {
        "type": "object",
        "required": [
          "message",
          "data"
        ],
        "properties": {
          "data": {
            "description": "响应数据",
            "x-go-name": "Data"
          },
          "message": {
            "description": "响应消息",
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    },
...
  }
}